requirements.txt
python-binance>=1.0.15
pandas>=1.5.0
numpy>=1.21.0
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
ta>=0.9.0
plotly>=5.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
websocket-client>=1.0.0
psutil>=5.8.0